Updated tests to check against uphold verified Refactored tests and controllers to be more RESTful
Closes #1078
